expression: improve code to split base interface
