+++ expression/internal/base.go
@@ -1,0 +1,33 @@
+package internal
+
+import (
+	"github.com/pingcap/tidb/sessionctx"
+	"github.com/pingcap/tidb/util/chunk"
+)
+
+// VecExpr contains all vectorized evaluation methods.
+type VecExpr interface {
+	// Vectorized returns if this expression supports vectorized evaluation.
+	Vectorized() bool
+
+	// VecEvalInt evaluates this expression in a vectorized manner.
+	VecEvalInt(ctx sessionctx.Context, input *chunk.Chunk, result *chunk.Column) error
+
+	// VecEvalReal evaluates this expression in a vectorized manner.
+	VecEvalReal(ctx sessionctx.Context, input *chunk.Chunk, result *chunk.Column) error
+
+	// VecEvalString evaluates this expression in a vectorized manner.
+	VecEvalString(ctx sessionctx.Context, input *chunk.Chunk, result *chunk.Column) error
+
+	// VecEvalDecimal evaluates this expression in a vectorized manner.
+	VecEvalDecimal(ctx sessionctx.Context, input *chunk.Chunk, result *chunk.Column) error
+
+	// VecEvalTime evaluates this expression in a vectorized manner.
+	VecEvalTime(ctx sessionctx.Context, input *chunk.Chunk, result *chunk.Column) error
+
+	// VecEvalDuration evaluates this expression in a vectorized manner.
+	VecEvalDuration(ctx sessionctx.Context, input *chunk.Chunk, result *chunk.Column) error
+
+	// VecEvalJSON evaluates this expression in a vectorized manner.
+	VecEvalJSON(ctx sessionctx.Context, input *chunk.Chunk, result *chunk.Column) error
+}
